airflow/requirement.txt
pandas
boto3
scikit-learn
joblib
python-dotenv
Python-IO
pymongo
pymongo[srv]
prophet
dnspython